Playtest UI and feedback changes
Lots of hacks into the UI to make it obvious that it's a PC playtest server.  Added functionality to put messages in the chat with links to the feedback form, and to repeat that information in a pop-up.  Due to the nature of these changes, they can't be tested until they're loaded on an actual server, so here we are.